fix: Force close and join at exit
